+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Added search functionality on jobs table filters (\#424).
+
 # 0.10.1
 
 * Fixed bug in workflow task meta properties update (\#422).